refactor card-type to be a service
