+++ packages/host/app/components/operator-mode/card-adoption-chain.gts
@@ -11,9 +11,10 @@ import { CardInheritance } from '@cardstack/host/components/operator-mode/code-s
 import { Divider } from '@cardstack/host/components/operator-mode/definition-container';
 
 import { stripFileExtension } from '@cardstack/host/lib/utils';
-import { Type } from '@cardstack/host/resources/card-type';
+
 import type { Ready } from '@cardstack/host/resources/file';
 
+import type { Type } from '@cardstack/host/services/card-type';
 import { isOwnField } from '@cardstack/host/utils/schema-editor';
 
 interface Signature {
