+++ packages/host/app/components/operator-mode/code-submode/schema-editor.gts
@@ -10,11 +10,13 @@ import { type CodeRef } from '@cardstack/runtime-common/code-ref';
 import { ModuleSyntax } from '@cardstack/runtime-common/module-syntax';
 
 import CardAdoptionChain from '@cardstack/host/components/operator-mode/card-adoption-chain';
-import { CardType, Type } from '@cardstack/host/resources/card-type';
 import { Ready } from '@cardstack/host/resources/file';
 import { inheritanceChain } from '@cardstack/host/resources/inheritance-chain';
 import type { ModuleAnalysis } from '@cardstack/host/resources/module-contents';
+import type { Type } from '@cardstack/host/services/card-type';
+import { type CardType } from '@cardstack/host/services/card-type';
 import type LoaderService from '@cardstack/host/services/loader-service';
+
 import { calculateTotalOwnFields } from '@cardstack/host/utils/schema-editor';
 
 import { BaseDef } from 'https://cardstack.com/base/card-api';
