+++ packages/host/app/utils/schema-editor.ts
@@ -1,4 +1,4 @@
-import { type Type } from '@cardstack/host/resources/card-type';
+import { type Type } from '@cardstack/host/services/card-type';
 
 import type { BaseDef } from 'https://cardstack.com/base/card-api';
